+++ packages/desktop-client/src/util/router-tools.ts
@@ -9,7 +9,3 @@ export function ExposeNavigate() {
   }, [navigate]);
   return null;
 }
-
-export function openUrl(url: string) {
-  window.open(url, '_blank', 'noopener,noreferrer');
-}
