+++ upcoming-release-notes/3699.md
@@ -1,0 +1,6 @@
+---
+category: Enhancements
+authors: [jfdoming]
+---
+
+Update Electron help menu to reflect new in-app menu
